--- app/src/main/java/com/moroom/android/ReviewAdapter.kt
@@ -1,35 +1,0 @@
-package com.moroom.android
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import com.moroom.android.databinding.ListItemBinding
-
-class ReviewAdapter(private val reviewList: ArrayList<Review>) :
-    RecyclerView.Adapter<ReviewAdapter.ReviewViewHolder>() {
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ReviewViewHolder {
-        return ReviewViewHolder.from(parent)
-    }
-
-    override fun onBindViewHolder(holder: ReviewViewHolder, position: Int) {
-        holder.bind(reviewList[position])
-    }
-
-    override fun getItemCount(): Int = reviewList.size
-
-    class ReviewViewHolder private constructor(val binding: ListItemBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(item: Review) {
-            binding.item = item
-        }
-
-        companion object {
-            fun from(parent: ViewGroup): ReviewViewHolder {
-                val layoutInflater = LayoutInflater.from(parent.context)
-                val binding = ListItemBinding.inflate(layoutInflater, parent, false)
-                return ReviewViewHolder(binding)
-            }
-        }
-    }
-}
